pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marco Insabato <marcoins95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.54.4"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
spotipy = "^2.24.0"

[tool.poetry.scripts] 
say_hello = "app.commands.say_hello:main" 
serve = "app.commands.serve:main" 
make = "app.commands.make:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
